+++ packages/commercetools/theme/helpers/README.md
@@ -1,0 +1,1 @@
+Put here platform-specific, non-agnostic functions that overwrite default code.
