+++ packages/commercetools/theme/helpers/category/getCategorySearchParameters.js
