--- packages/commercetools/theme/includes/README.md
@@ -1,1 +1,0 @@
-Put here platform-specific, non-agnostic functions that overwrite default code. Use only as a last resort solution.
