+++ packages/core/theme-module/index.js
@@ -21,7 +21,7 @@ module.exports = async function DefaultThemeModule(moduleOptions) {
   const projectLocalThemeDir = this.options.buildDir.replace('.nuxt', '.theme');
   const themeComponentsDir = path.join(this.options.rootDir, 'pages');
   const themePagesDir = path.join(this.options.rootDir, 'components');
-  const themeIncludesDir = path.join(this.options.rootDir, 'includes');
+  const themeHelpersDir = path.join(this.options.rootDir, 'helpers');
   const themeFiles = getAllFilesFromDir(baseThemeDir).filter(file => !file.includes('/static/'));
 
   const compileAgnosticTemplate = (filePath) => {
@@ -41,7 +41,7 @@ module.exports = async function DefaultThemeModule(moduleOptions) {
   await Promise.all([
     copyThemeFiles(themeComponentsDir),
     copyThemeFiles(themePagesDir),
-    copyThemeFiles(themeIncludesDir)
+    copyThemeFiles(themeHelpersDir)
   ]);
 
   log.success(`Added ${themeFiles.length} theme file(s) to ${chalk.bold('.theme')} folder`);
@@ -139,7 +139,7 @@ module.exports = async function DefaultThemeModule(moduleOptions) {
       }
     });
 
-    chokidar.watch([themeComponentsDir, themePagesDir, themeIncludesDir], { ignoreInitial: true })
+    chokidar.watch([themeComponentsDir, themePagesDir, themeHelpersDir], { ignoreInitial: true })
       .on('all', (event, filePath) => {
         if (event === 'unlink') {
           const baseFilePath = filePath.replace(this.options.rootDir, baseThemeDir);
