+++ packages/core/theme-module/theme/helpers/category/getCategorySearchParameters.js
