+++ packages/core/theme-module/theme/pages/Category.vue
@@ -310,7 +310,7 @@ import {
 } from '@storefront-ui/vue';
 import { computed, ref, watch } from '@vue/composition-api';
 import { useCategory, useProduct, productGetters, categoryGetters } from '<%= options.composables %>';
-import { getCategorySearchParameters } from '~/includes/getCategorySearchParameters';
+import { getCategorySearchParameters } from '~/helpers/category/getCategorySearchParameters';
 import { onSSR } from '@vue-storefront/core';
 
 const perPageOptions = [20, 40, 100];
